src/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.2
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
isort==5.13.2
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
passlib==1.7.4
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
websockets==12.0
